Add the LimitSelect function to SelectBuilder (#216)
In pgsql it is allowed to do a SELECT statement for a LIMIT, as follows:

	SELECT * FROM foo
	WHERE ...
	LIMIT (SELECT col FROM bar)

This commit adds a function `LimitSelect, similar to the FromSelect function, that allows for
a SelectBuilder to be used for to create the limit clause.

This commit also adds a utility function `Parenthesized`, that simply returns a `Sqlizer` that
adds parentheses around the expression.